+++ RE4_UHD_BIN_TOOL/SHARED_UHD_BIN/REPACK/Structures/FinalStructure.cs
@@ -36,15 +36,15 @@ namespace SHARED_UHD_BIN.REPACK.Structures
 
     public class FinalWeightMap : IEquatable<FinalWeightMap>
     {
-        public ushort Links { get; set; }
+        public byte Links { get; set; }
 
-        public ushort BoneID1 { get; set; }
+        public byte BoneID1 { get; set; }
         public byte Weight1 { get; set; }
 
-        public ushort BoneID2 { get; set; }
+        public byte BoneID2 { get; set; }
         public byte Weight2 { get; set; }
 
-        public ushort BoneID3 { get; set; }
+        public byte BoneID3 { get; set; }
         public byte Weight3 { get; set; }
 
         public override bool Equals(object obj)
